Finished the Nav Bar hovering action.
Made the insets in nav_bar.css smaller.

Added nav_bar.css to the Scene in MainWindow.java.

Swapped the sides of the home button and other buttons in nav_bar.fxml

Changed the expanding of the Nav Bar to occur on hovering rather than clicking of navigationButton in NavBarController.java