Code Refactoring and Search Screen Styling:
Refactored SearchScreenController.java:
Created a service class for the controller: SearchScreenService.java
Cleaned up the initialize method to be more readable/maintainable
Changed an if assignment to use a ternary operator

Styling:
Changed the fillVbox method to generate the buttons using a service method in the service class that styles them and adds an appropriate image based on the colour of the wine they refer to.

Added temporary images to the images directory for the wine buttons.