Fixed nav bar expansion and compression:
Changes to NavBarController.java:
Edited the KeyValues of the animations in expandNavBar to take into account the expansion of the images.

Edited the KeyValues of the animation in closeNavBar to be the opposite of those in expandNavBar.

Changes to NavBarService.java:
Altered the two methods to return an ImageView so that it can be used by expandNavBar and closeNavBar in NavBarController.java