Added (Actually Informative) text into the help screen:
Created a new resources directory: text. Holds all the text required by the help screen content area so that it is easily editable if it needs to be changed.

Added three files into that directory: sign_in.txt, about_winedy.txt, and create_an_account.txt. These all contain useful information about how to execute the relevant use cases.

Created a new guiservice class: HelpScreenService.java which currently contains one method getContentFromFile which reads in the text from the given file and returns a String to be used as the text in the labels in HelpScreenController.java.