Quick fixes before deliverable:
Removed the assigning of the database to a variable in App.java, the getInstance method is only called.

Fixed the adding of the CSS style sheets in MainWindow.java. There was a double up in one of the paths and red_wine_scroll_pane.css is missing.

Commented out the line of code adding a style to the combo box in SearchScreenController.java. Also fixed some convention inconsistencies.