Added Winedy Name To The NavBar:
Edited nav_bar.fxml to have a new rectangle and image view in the middle of the nav bar. Injected these into NavBarController.java and edited the initialize method to set their style up.

Added winedy_text.png - did this rather than using text/a label as I wanted the 'y' to be a different colour from the 'Wined'.