Added a check when creating wineList buttons in WineListSelectPopUp so the favourites list is excluded.
Updated FXWrapper to have a previous screen stack. You can push a screen to the stack and calling the loadPreviousScreen method will pop the latest screen and load it. This can be used dynamically when multiple pop ups are stacked on top of each other.

Updated ProfileWineListsScreenController and CreateNewListPopUpController to use the new previous stack functionality.

The add wine to list button now allows user to add a new list easily without having to go back to the profile screen.