Started styling the search screen.
Removed everything from the gridpanes.

Increased the overall size of all the widgets.

Moved all the filters to the right side of the page.

Added a rectangle behind the search results scroll pane with the red-wine-rectangle style class.

Added the nav-bar-button style class to the buttons in the vbox of the search results.