Continued styling the search screen
Put all of the filters into a vbox that gets expanded and compressed by the on action of a button.

Added rectangles behind the groups of widgets and styled them

Added an information message when the screen is first opened and changed the message that displays when the search returns nothing

Slightly changed the generation of the wine buttons so that they have a larger font.

Coded the logic in SearchScreenController.java for the expansion and compression of the filters vbox