Finished extracting the Wine List view from the search screen.
Removed the original vbox and scroll pane from search_screen.fxml and replaced them with the Anchor pane for the view to be loaded into.

Finished up the WineListViewController.java and linked it to the FXML file.

Altered the add image graphic to button method to handle the exceptional case where the image path is invalid, it now displays a text graphic showing that the image was not found.

Made it so that the search screen automatically loads all the wines on the first instance of it being open, and stores the most recent search in FXWrapper so it can be reloaded on initialisation.