Made the Search Screen load the previous search on reopening
Added the search attributes into SearchWineListTest.java and redefined its constructor to take all the attributes of the search.
(This constructor is protected against null inputs)

Had to refactor SearchStepDefinitions.java, SearchStepDefinitions.java, SearchWineListTest.java, WineDAO.java, and WineListTest.java to use the new constructor.

When the search screen is opened, if a previous search is stored in WineListManager, this is loaded into the screen.

Removed the previous searched functionality from FXWrapper as this was already defined in WineListManager which is a more appropriate place to store it.

ISSUES: Currently there are issues with going back to the search screen with all in the search