Updated ProfileWineListsScreenController so the delete mode works properly. Had some issues with updating pages when changing to/from delete mode.
Changed deleteMode from boolean to BooleanProperty and added a listener to all buttons. This way buttons update whenever delete mode is toggled.

Simplified pagination and removed pageCache. Since the page content needs to reflect changes in current state of delete mode and wine lists, caching does not seem necessary.

(Detailed Explanation) There is an issue with pagination.setCurrentPageIndex() where it does not reload the page if you pass in the same index as current page index. This means we can't use this method to reload current page whenever delete mode is toggled. So instead I added listeners to the buttons so that they can update themselves when delete mode is toggled.