Redesigned the individual wine view popup
The pop-up is now larger to accommodate all the text to be displayed.

Made the title label larger to allow the largest wine name to be displayed without extending over other parts of the popup.

Added an image in the middle of the popup to convey the colour of the wine.

Moved the awards to the right of the image and made the description fit the entire width of the popup below the image.

Fixed the issue where the style and country labels were not being hidden if the wine did not have these attributes (The condition was testing for null rather than the empty string)