Increased the size of the font in the help screen to be a bit more readable
Made the contents labels be underlined to indicate they are intractable.

Added a label with an instruction to indicate this too

Decreased the size of the contents scroll pane to separate it from the instruction label.

Fixed a bug in the setContentsButtonOnAction where the viewPortHeight was getting it from the wrong scroll pane.